fix uplode ecxle

+++ database/migrations/2025_04_26_131646_add_upload_id_to_students_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up()
     {
         Schema::table('students', function (Blueprint $table) {
-            $table->integer('upload_id')->nullable()->after('sujet');
+            $table->bigInteger('upload_id')->nullable()->after('sujet');
         });
     }
